global: make openssl registration like the rest
